+++ src/components/atoms/vCustomAutocomplete.vue
@@ -1,0 +1,22 @@
+<template>
+	<v-autocomplete
+		small-chips
+		deletable-chips
+		multiple
+		clearable
+		solo
+		rounded
+		:menu-props="{ closeOnContentClick: false }"
+		v-bind="$attrs"
+		@change="($event) => $emit('selected-item', $event)"
+	/>
+</template>
+
+<script>
+export default {
+	model: {
+		prop: "value",
+		event: "selected-item",
+	},
+};
+</script>