+++ src/store/homeworks.unit.js
@@ -1,7 +1,5 @@
 import {
 	homeworks,
-	openHomeworks,
-	completedHomeworks,
 	submittedHomeworks,
 	gradedHomeworks,
 	openHomeworksWithDueDate,
@@ -9,8 +7,6 @@ import {
 	overDueHomeworks,
 	coursesStudent,
 	coursesTeacher,
-	coursesOpen,
-	coursesCompleted,
 	mathHomeworks,
 	homeworksTeacher,
 	overDueHomeworksTeacher,
@@ -18,29 +14,16 @@ import {
 	dueDateHomeworksTeacher,
 } from "@@/stories/mockData/Homeworks";
 import storeModule from "./homeworks";
-import AuthModule from "@/store/auth";
 
 describe("store/homeworks", () => {
 	describe("actions", () => {
 		const spyCommit = jest.fn();
-		const ctxMockTeacher = {
+		const ctxMock = {
 			commit: spyCommit,
-			rootState: {
-				auth: { user: { permissions: ["TASK_DASHBOARD_TEACHER_VIEW_V3"] } },
-			},
 		};
 
-		const ctxMockStudent = {
-			commit: spyCommit,
-			rootState: {
-				auth: { user: { permissions: ["TASK_DASHBOARD_VIEW_V3"] } },
-			},
-		};
-
-		describe("getHomeworksDashboard", () => {
-			it("should call the right endpoint for teachers", async () => {
-				AuthModule.setUser(ctxMockTeacher.rootState.auth.user);
-
+		describe("getAllHomeworks", () => {
+			it("should call the right endpoint", async () => {
 				const receivedRequests = [];
 
 				storeModule.actions.$axios = {
@@ -50,31 +33,10 @@ describe("store/homeworks", () => {
 					},
 				};
 
-				await storeModule.actions.getHomeworksDashboard(ctxMockTeacher);
-				expect(spyCommit.mock.calls).toHaveLength(3);
-				expect(spyCommit.mock.calls[1][0]).toBe("set");
-				expect(receivedRequests[0].url).toStrictEqual("/v3/tasks/open/");
-			});
-
-			it("should call both endpoint for students", async () => {
-				AuthModule.setUser(ctxMockStudent.rootState.auth.user);
-
-				const receivedRequests = [];
-
-				storeModule.actions.$axios = {
-					$get: async (url, params) => {
-						receivedRequests.push({ url, params });
-						return { data: [] };
-					},
-				};
-
-				spyCommit.mockClear();
-				await storeModule.actions.getHomeworksDashboard(ctxMockStudent);
-
-				expect(spyCommit.mock.calls).toHaveLength(3);
-				expect(spyCommit.mock.calls[1][0]).toBe("set");
-				expect(receivedRequests[0].url).toStrictEqual("/v3/tasks/open/");
-				expect(receivedRequests[1].url).toStrictEqual("/v3/tasks/completed/");
+				await storeModule.actions.getAllHomeworks(ctxMock);
+				expect(spyCommit.mock.calls).toHaveLength(4);
+				expect(spyCommit.mock.calls[2][0]).toBe("setHomeworks");
+				expect(receivedRequests[0].url).toStrictEqual("/v3/tasks/");
 			});
 
 			it("should set business error and reset loading on fail", async () => {
@@ -94,24 +56,26 @@ describe("store/homeworks", () => {
 				};
 
 				spyCommit.mockClear();
-				await storeModule.actions.getHomeworksDashboard(ctxMockTeacher);
+				await storeModule.actions.getAllHomeworks(ctxMock);
 
 				const storeCalls = spyCommit.mock.calls;
 				const firstCall = storeCalls[0];
 				const secondCall = storeCalls[1];
 				const thirdCall = storeCalls[2];
+				const fourthCall = storeCalls[3];
 
-				expect(firstCall[0]).toBe("setStatus");
-				expect(secondCall[0]).toBe("setBusinessError");
-				expect(thirdCall[0]).toBe("setStatus");
+				expect(firstCall[0]).toBe("resetBusinessError");
+				expect(secondCall[0]).toBe("setStatus");
+				expect(thirdCall[0]).toBe("setBusinessError");
+				expect(fourthCall[0]).toBe("setStatus");
 
-				const firstCommit = firstCall[1];
 				const secondCommit = secondCall[1];
 				const thirdCommit = thirdCall[1];
+				const fourthCommit = fourthCall[1];
 
-				expect(firstCommit).toBe("pending");
-				expect(secondCommit).toBe(mockAxiosError.response.data);
-				expect(thirdCommit).toBe("error");
+				expect(secondCommit).toBe("pending");
+				expect(thirdCommit).toBe(mockAxiosError.response.data);
+				expect(fourthCommit).toBe("error");
 			});
 
 			it("should set loading state when fetching homeworks", async () => {
@@ -124,26 +88,37 @@ describe("store/homeworks", () => {
 					},
 				};
 				spyCommit.mockClear();
-				await storeModule.actions.getHomeworksDashboard(ctxMockTeacher);
+				await storeModule.actions.getAllHomeworks(ctxMock);
 
 				const storeCalls = spyCommit.mock.calls;
 
-				const firstCall = storeCalls[0];
-				const thirdCall = storeCalls[2];
-				expect(firstCall[0]).toBe("setStatus");
-				expect(thirdCall[0]).toBe("setStatus");
+				const secondCall = storeCalls[1];
+				const fourthCall = storeCalls[3];
+				expect(secondCall[0]).toBe("setStatus");
+				expect(fourthCall[0]).toBe("setStatus");
 
-				const firstCommit = firstCall[1];
-				const secondCommit = thirdCall[1];
-				expect(firstCommit).toBe("pending");
-				expect(secondCommit).toBe("completed");
+				const secondCommit = secondCall[1];
+				const fourthCommit = fourthCall[1];
+				expect(secondCommit).toBe("pending");
+				expect(fourthCommit).toBe("completed");
 			});
 		});
 	});
 
 	describe("mutations", () => {
+		describe("setHomeworks", () => {
+			it("Should set homeworks in state", () => {
+				const { setHomeworks } = storeModule.mutations;
+				const state = {
+					homeworks: [],
+				};
+				setHomeworks(state, homeworks);
+				expect(state.homeworks).toStrictEqual(homeworks);
+			});
+		});
+
 		describe("setFilter", () => {
-			it("Should set course filter in the state", () => {
+			it("Should set course filter in state", () => {
 				const { setFilter } = storeModule.mutations;
 				const state = {
 					courseFilter: [],
@@ -152,213 +127,205 @@ describe("store/homeworks", () => {
 				expect(state.courseFilter).toStrictEqual(coursesTeacher);
 			});
 		});
+
+		describe("setStatus", () => {
+			it("Should set status in state", () => {
+				const { setStatus } = storeModule.mutations;
+				const state = {
+					status: "",
+				};
+				setStatus(state, "pending");
+				expect(state.status).toStrictEqual("pending");
+			});
+		});
+
+		describe("setBusinessError", () => {
+			it("Should set business error in state", () => {
+				const { setBusinessError } = storeModule.mutations;
+				const state = {
+					businessError: {},
+				};
+				setBusinessError(state, {
+					statusCode: "404",
+					message: "404 - Can't find page",
+				});
+				expect(state.businessError).toStrictEqual({
+					statusCode: "404",
+					message: "404 - Can't find page",
+				});
+			});
+		});
+
+		describe("resetBusinessError", () => {
+			it("Should reset business error in state", () => {
+				const { resetBusinessError } = storeModule.mutations;
+				const state = {
+					businessError: {},
+				};
+				resetBusinessError(state);
+				expect(state.businessError).toStrictEqual({
+					statusCode: "",
+					message: "",
+				});
+			});
+		});
 	});
 
 	describe("getters", () => {
-		const state = {
-			list: homeworks,
+		const studentState = {
+			homeworks,
 			status: "completed",
 			courseFilter: [],
 		};
-		const stateWithFilter = {
-			list: homeworksTeacher,
+		const teacherState = {
+			homeworks: homeworksTeacher,
+			status: "completed",
+			courseFilter: [],
+		};
+		const teacherStateWithFilter = {
+			homeworks: homeworksTeacher,
 			status: "completed",
 			courseFilter: ["Mathe"],
 		};
+		const emptyState = {
+			homeworks: [],
+			status: "completed",
+			courseFilter: [],
+		};
 		const { getters } = storeModule;
 
-		describe("isListEmpty", () => {
-			it("Should return true, if it's loaded and there are no homeworks", () => {
-				expect(getters.isListEmpty(state)).toBe(false);
-			});
-		});
-
-		describe("isListFilled", () => {
-			it("Should return true, if it's loaded and there are homeworks", () => {
-				expect(getters.isListFilled(state)).toBe(true);
-			});
-		});
-
-		describe("getCourses", () => {
-			it("Should return all relevant courses", () => {
-				expect(getters.getCourses(state)).toStrictEqual(coursesStudent);
-			});
-		});
-
-		describe("getCoursesOpen", () => {
-			it("Should return all relevant open courses", () => {
-				const mockGetter = {
-					getOpenHomeworks: openHomeworks,
-				};
-				expect(getters.getCoursesOpen(state, mockGetter)).toStrictEqual(
-					coursesOpen
-				);
-			});
-		});
-
-		describe("getCoursesCompleted", () => {
-			it("Should return all relevant completed courses", () => {
-				const mockGetter = {
-					getCompletedHomeworks: completedHomeworks,
-				};
-				expect(getters.getCoursesCompleted(state, mockGetter)).toStrictEqual(
-					coursesCompleted
-				);
-			});
-		});
-
-		describe("getOpenHomeworksWithDueDate", () => {
-			it("Should return homeworks before due date", () => {
-				const mockGetter = {
-					getOpenHomeworks: openHomeworks,
-				};
-				expect(
-					getters.getOpenHomeworksWithDueDate(state, mockGetter)
-				).toHaveLength(openHomeworksWithDueDate.length);
-			});
-		});
-
-		describe("getOpenHomeworksWithDueDateTeacher", () => {
-			it("Should return homeworks before due date for teachers", () => {
-				const mockGetter = {
-					getHomeworks: homeworksTeacher,
-				};
-				expect(
-					getters.getOpenHomeworksWithDueDateTeacher(state, mockGetter)
-				).toHaveLength(dueDateHomeworksTeacher.length);
-			});
-		});
-
-		describe("getOverDueHomeworks", () => {
-			it("Should return homeworks after due date", () => {
-				const mockGetter = {
-					getOpenHomeworks: openHomeworks,
-				};
-				expect(getters.getOverDueHomeworks(state, mockGetter)).toHaveLength(
-					overDueHomeworks.length
-				);
-			});
-		});
-
-		describe("getOverDueHomeworksTeacher", () => {
-			it("Should return homeworks for teachers after due date", () => {
-				const mockGetter = {
-					getHomeworks: homeworksTeacher,
-				};
-				expect(
-					getters.getOverDueHomeworksTeacher(state, mockGetter)
-				).toHaveLength(overDueHomeworksTeacher.length);
-			});
-		});
-
-		describe("getOpenHomeworksWithoutDueDate", () => {
-			it("Should return open homeworks without due date", () => {
-				const mockGetter = {
-					getOpenHomeworks: openHomeworks,
-				};
-				expect(
-					getters.getOpenHomeworksWithoutDueDate(state, mockGetter)
-				).toHaveLength(openHomeworksWithoutDueDate.length);
-			});
-		});
-
-		describe("getOpenHomeworksWithoutDueDateTeacher", () => {
-			it("Should return open homeworks without due date for teachers", () => {
-				const mockGetter = {
-					getHomeworks: homeworksTeacher,
-				};
-				expect(
-					getters.getOpenHomeworksWithoutDueDateTeacher(state, mockGetter)
-				).toHaveLength(noDueDateHomeworksTeacher.length);
-			});
-		});
-
 		describe("getHomeworks", () => {
-			it("Should return filtered homeworks, if filter is applied", () => {
-				const mockGetter = { getFilteredHomeworks: mathHomeworks };
-
-				expect(getters.getHomeworks(stateWithFilter, mockGetter)).toStrictEqual(
-					mathHomeworks
-				);
-			});
-
-			it("Should return all homeworks, if no filter is applied", () => {
-				const mockGetter = { getFilteredHomeworks: [] };
-				expect(getters.getHomeworks(state, mockGetter)).toStrictEqual(
-					homeworks
-				);
+			it("Should return all homeworks", () => {
+				expect(getters.getHomeworks(studentState)).toStrictEqual(homeworks);
 			});
 		});
 
-		describe("getOpenHomeworks", () => {
-			it("Should return all homeworks, that neither submitted nor graded", () => {
-				const mockGetter = { getHomeworks: homeworks };
-				expect(getters.getOpenHomeworks(state, mockGetter)).toStrictEqual(
-					openHomeworks
-				);
+		describe("getStatus", () => {
+			it("Should return status", () => {
+				expect(getters.getStatus(studentState)).toStrictEqual("completed");
 			});
 		});
 
-		describe("getCompletedHomeworks", () => {
-			it("Should return all homeworks, that are either submitted or graded", () => {
-				const mockGetter = { getHomeworks: homeworks };
-				expect(getters.getCompletedHomeworks(state, mockGetter)).toStrictEqual(
-					completedHomeworks
-				);
-			});
-		});
-
-		describe("getSubmittedHomeworks", () => {
-			it("Should return all homeworks, that are submitted", () => {
-				const mockGetter = { getHomeworks: homeworks };
-				expect(getters.getSubmittedHomeworks(state, mockGetter)).toStrictEqual(
-					submittedHomeworks
-				);
-			});
-		});
-
-		describe("getGradedHomeworks", () => {
-			it("Should return all homeworks, that are graded", () => {
-				const mockGetter = { getCompletedHomeworks: completedHomeworks };
-				expect(getters.getGradedHomeworks(state, mockGetter)).toStrictEqual(
-					gradedHomeworks
-				);
-			});
-		});
-
-		describe("getFilteredHomeworks", () => {
-			it("Should return homeworks filtered by course", () => {
-				expect(getters.getFilteredHomeworks(stateWithFilter)).toStrictEqual(
-					mathHomeworks
-				);
-			});
-		});
-
-		describe("hasOpenHomeworks", () => {
-			it("Should return true, if it's loaded and there are open (neither submitted nor graded) homeworks", () => {
-				const mockGetter = { getOpenHomeworks: openHomeworks };
-				expect(getters.hasOpenHomeworks(state, mockGetter)).toBe(true);
+		describe("hasNoHomeworks", () => {
+			it("Should return true, if it's loaded and there are no homeworks", () => {
+				expect(getters.hasNoHomeworks(studentState)).toBe(false);
 			});
 		});
 
 		describe("hasNoOpenHomeworks", () => {
-			it("Should return true, if it's loaded and there are no open (neither submitted nor graded) homeworks", () => {
-				const mockGetter = { getOpenHomeworks: [] };
-				expect(getters.hasNoOpenHomeworks(state, mockGetter)).toBe(true);
-			});
-		});
-
-		describe("hasCompletedHomeworks", () => {
-			it("Should return true, if it's loaded and there are completed homeworks", () => {
-				const mockGetter = { getCompletedHomeworks: completedHomeworks };
-				expect(getters.hasCompletedHomeworks(state, mockGetter)).toBe(true);
+			it("Should return true, if it's loaded and there are no open homeworks", () => {
+				expect(getters.hasNoOpenHomeworks(emptyState)).toBe(true);
 			});
 		});
 
 		describe("hasNoCompletedHomeworks", () => {
 			it("Should return true, if it's loaded and there are no completed homeworks", () => {
-				const mockGetter = { getCompletedHomeworks: [] };
-				expect(getters.hasNoCompletedHomeworks(state, mockGetter)).toBe(true);
+				expect(getters.hasNoCompletedHomeworks(emptyState)).toBe(true);
+			});
+		});
+
+		describe("getOpenHomeworksForStudent", () => {
+			it("Should have properties for sub sets", () => {
+				expect(getters.getOpenHomeworksForStudent(studentState)).toHaveProperty(
+					"overdue"
+				);
+				expect(getters.getOpenHomeworksForStudent(studentState)).toHaveProperty(
+					"noDueDate"
+				);
+				expect(getters.getOpenHomeworksForStudent(studentState)).toHaveProperty(
+					"withDueDate"
+				);
+			});
+
+			it("Should have homeworks with due date", () => {
+				expect(
+					getters.getOpenHomeworksForStudent(studentState).withDueDate
+				).toHaveLength(openHomeworksWithDueDate.length);
+			});
+
+			it("Should have homeworks with no due date", () => {
+				expect(
+					getters.getOpenHomeworksForStudent(studentState).noDueDate
+				).toHaveLength(openHomeworksWithoutDueDate.length);
+			});
+
+			it("Should have overdue homeworks", () => {
+				expect(
+					getters.getOpenHomeworksForStudent(studentState).overdue
+				).toHaveLength(overDueHomeworks.length);
+			});
+		});
+
+		describe("getOpenHomeworksForTeacher", () => {
+			it("Should have properties for sub sets", () => {
+				expect(getters.getOpenHomeworksForTeacher(teacherState)).toHaveProperty(
+					"overdue"
+				);
+				expect(getters.getOpenHomeworksForTeacher(teacherState)).toHaveProperty(
+					"noDueDate"
+				);
+				expect(getters.getOpenHomeworksForTeacher(teacherState)).toHaveProperty(
+					"withDueDate"
+				);
+			});
+
+			it("Should have homeworks with due date", () => {
+				expect(
+					getters.getOpenHomeworksForTeacher(teacherState).withDueDate
+				).toHaveLength(dueDateHomeworksTeacher.length);
+			});
+
+			it("Should have homeworks with no due date", () => {
+				expect(
+					getters.getOpenHomeworksForTeacher(teacherState).noDueDate
+				).toHaveLength(noDueDateHomeworksTeacher.length);
+			});
+
+			it("Should have overdue homeworks", () => {
+				expect(
+					getters.getOpenHomeworksForTeacher(teacherState).overdue
+				).toHaveLength(overDueHomeworksTeacher.length);
+			});
+
+			it("Should correctly filter homeworks", () => {
+				const filteredHomeworks = getters.getOpenHomeworksForTeacher(
+					teacherStateWithFilter
+				);
+
+				expect(
+					filteredHomeworks.overdue
+						.concat(filteredHomeworks.noDueDate)
+						.concat(filteredHomeworks.withDueDate)
+				).toHaveLength(mathHomeworks.length);
+			});
+		});
+
+		describe("getCompletedHomeworksForStudent", () => {
+			it("Should have properties for sub sets", () => {
+				expect(
+					getters.getCompletedHomeworksForStudent(studentState)
+				).toHaveProperty("submitted");
+				expect(
+					getters.getCompletedHomeworksForStudent(studentState)
+				).toHaveProperty("graded");
+			});
+
+			it("Should have all homeworks, that are submitted", () => {
+				expect(
+					getters.getCompletedHomeworksForStudent(studentState).submitted
+				).toStrictEqual(submittedHomeworks);
+			});
+
+			it("Should have all homeworks, that are graded", () => {
+				expect(
+					getters.getCompletedHomeworksForStudent(studentState).graded
+				).toStrictEqual(gradedHomeworks);
+			});
+		});
+
+		describe("getCourses", () => {
+			it("Should return all relevant courses", () => {
+				expect(getters.getCourses(studentState)).toStrictEqual(coursesStudent);
 			});
 		});
 	});